removed BasePaths, moved ClientRoutes into meta
