+++ src/client/components/CountrySelect/CountryList/CountryListRow/CountryListRow.tsx
@@ -2,12 +2,12 @@ import React, { useEffect, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Link } from 'react-router-dom'
 
+import { ClientRoutes } from '@meta/app'
 import { Areas, CountryIso, Global, RegionCode } from '@meta/area'
 import { UserRoles } from '@meta/user/userRoles'
 
 import { useAssessment, useCountry, useCycle } from '@client/store/assessment'
 import { useCountryIso, useIsHome } from '@client/hooks'
-import { ClientRoutes } from '@client/clientRoutes'
 import { Dates } from '@client/utils'
 
 type Props = {
@@ -37,6 +37,7 @@ const CountryListRow: React.FC<Props> = (props: Props) => {
     if (selected) {
       countryNameRef.current.scrollIntoView({ behavior: 'smooth', block: 'center', inline: 'nearest' })
     }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
   return (
@@ -56,7 +57,7 @@ const CountryListRow: React.FC<Props> = (props: Props) => {
         <>
           <span className="country-selection-list__secondary-col">
             <div className={`status-${status}`} />
-            {i18n.t(`assessment.status.${status}.label`)}
+            {i18n.t<string>(`assessment.status.${status}.label`)}
           </span>
 
           <span className="country-selection-list__secondary-col">
