+++ src/client/components/CountrySelect/LinkLanding/LinkLanding.tsx
@@ -5,9 +5,10 @@ import { matchPath, NavLink, useLocation } from 'react-router-dom'
 
 import classNames from 'classnames'
 
+import { ClientRoutes } from '@meta/app'
+
 import { useAssessment, useCycle } from '@client/store/assessment'
 import { useCountryIso } from '@client/hooks'
-import { ClientRoutes } from '@client/clientRoutes'
 import Icon from '@client/components/Icon'
 
 const LinkLanding: React.FC = () => {
