+++ src/client/components/Navigation/NavigationDesktop/NavigationDesktop.tsx
@@ -2,13 +2,13 @@ import React, { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Link } from 'react-router-dom'
 
+import { ClientRoutes } from '@meta/app'
 import { Areas } from '@meta/area'
 
 import { useAppDispatch } from '@client/store'
 import { useAssessment, useCycle } from '@client/store/assessment'
 import { NavigationActions } from '@client/store/ui/navigation'
 import { useCountryIso } from '@client/hooks'
-import { ClientRoutes } from '@client/clientRoutes'
 import Icon from '@client/components/Icon'
 
 import NavAssessment from '../NavAssessment'
