+++ src/client/pages/AssessmentHome/FraHome/ButtonDownloadDashboard/ButtonDownloadDashboard.tsx
@@ -4,12 +4,11 @@ import { useTranslation } from 'react-i18next'
 import { Link, matchPath, useLocation } from 'react-router-dom'
 
 import { ApiEndPoint } from '@meta/api/endpoint'
-
+import { ClientRoutes } from '@meta/app'
 import { Areas } from '@meta/area'
 
 import { useAssessment, useCycle } from '@client/store/assessment'
 import { useCountryIso } from '@client/hooks'
-import { ClientRoutes } from '@client/clientRoutes'
 import Icon from '@client/components/Icon'
 
 const ButtonDownloadDashboard: React.FC = () => {
