+++ src/client/pages/OriginalDataPoint/components/YearSelection/YearSelection.tsx
@@ -5,13 +5,13 @@ import { useParams } from 'react-router-dom'
 import { Objects } from '@utils/objects'
 
 import { ApiEndPoint } from '@meta/api/endpoint'
+import { ClientRoutes } from '@meta/app'
 import { ODPYears, OriginalDataPoint } from '@meta/assessment'
 
 import { useAppDispatch } from '@client/store'
 import { useAssessment, useCycle } from '@client/store/assessment'
 import { OriginalDataPointActions, useOriginalDataPoint } from '@client/store/pages/originalDataPoint'
 import { useCountryIso, useGetRequest } from '@client/hooks'
-import { ClientRoutes } from '@client/clientRoutes'
 
 // TODO: Handle error
 const years = ['', ...ODPYears]
@@ -46,6 +46,7 @@ const YearSelection: React.FC<Props> = (props) => {
   )
   useEffect(() => {
     fetchReservedYears()
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [originalDataPoint.year])
 
   const reservedYears: Array<number> = data?.years ?? []
