--- src/meta/basePaths.ts
@@ -1,9 +1,0 @@
-export const BasePaths = {
-  Root: () => '/',
-
-  Login: {
-    root: () => '/login',
-    resetPassword: () => `/login/resetPassword`,
-    invitation: () => '/login/invitation',
-  },
-}
